requirements.txt
google-api-python-client
isodate
google-generativeai
transformers
torch
torchaudio
pydub
ffmpeg-python
